intermediate massive changes, not working yet
